Refactored production inventory related impl
