+++ src/utils/game/producing/inventory/fullPackTime.ts
@@ -1,0 +1,14 @@
+import {durationOfDay} from '@/const/common';
+
+
+type GetSecondsToFullPackInSleepOpts = {
+  dailyCount: number,
+  carryLimit: number,
+};
+
+export const getSecondsToFullPackInSleep = ({
+  dailyCount,
+  carryLimit,
+}: GetSecondsToFullPackInSleepOpts): number => {
+  return carryLimit / dailyCount * durationOfDay;
+};
